requirements.txt
Cairosvg==1.0.22
python-pdf==0.30.0
blinker==1.4
cairocffi==0.8
cffi==1.11.2
coverage==3.7.1
cssselect2==0.2.1
decorator==4.1.2
flask-login==0.3.0
flask-mail==0.9.0
flask-redis==0.3.0
flask-sqlalchemy==2.0
flask-socketio==3.0.2
flask-weasyprint==0.5
flask-whooshalchemy==0.56
flask-wtf==0.10.2
flask==0.10.1
flipflop==1.0
gunicorn==19.1.1
gevent==1.3.7
gevent-websocket==0.10.1
html5lib==0.999999999
itsdangerous==0.24
jinja2==2.10
markupsafe==0.23
pbr==0.10.0
pillow==5.4.1
psycopg2
pycparser==2.18
pyphen==0.9.4
pytz==2014.7
scikit-image==0.14.2
scipy==1.2.0
six==1.10.0
speaklater==1.3
sqlalchemy-migrate==0.9.2
sqlalchemy==0.9.7
sqlparse==0.1.11
tempita==0.5.2
tinycss2==0.6.0
weasyprint==0.41
webencodings==0.5.1
werkzeug==0.9.6
whoosh==2.6.0
wtforms==2.0.1